Start work on new onboarding flow
